Split Builder into Finder and Localizer
