+++ app_update_checker/lib/src/controller/update_contoller_base.dart
@@ -3,7 +3,7 @@
 import 'dart:async';
 
 import '../builder/models/app_update.dart';
-import '../builder/models/exceptions.dart';
+import 'exceptions.dart';
 import '../builder/models/release.dart';
 import '../builder/models/update_config.dart';
