+++ app_update_checker/lib/src/localizer/models/app_update.dart
