+++ app_update_checker/lib/src/localizer/models/release.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import '../../linker/models/release_data.dart';
 import '../../models/release_status.dart';
 import '../../models/text_translations.dart';
@@ -29,51 +27,6 @@ class Release extends ReleaseData {
     required super.stores,
     required super.customData,
   });
-
-  factory Release.localizedFromReleaseData({
-    required ReleaseData releaseData,
-    required Locale locale,
-    required String appName,
-    required Version appVersion,
-  }) {
-    String interpolation(String text) => text
-        .replaceAll(r'$appName', appName)
-        .replaceAll(
-          r'$appVersion',
-          appVersion.toString(),
-        )
-        .replaceAll(
-          r'$releaseVersion',
-          releaseData.version.toString(),
-        );
-
-    final title = interpolation(releaseData.titleTranslations.byLocale(locale));
-
-    final description = interpolation(releaseData.descriptionTranslations.byLocale(locale));
-
-    final releaseNoteTranslations = releaseData.releaseNoteTranslations;
-    final releaseNote =
-        releaseNoteTranslations == null ? null : interpolation(releaseNoteTranslations.byLocale(locale));
-
-    return Release(
-      version: releaseData.version,
-      refVersion: releaseData.refVersion,
-      buildNumber: releaseData.buildNumber,
-      status: releaseData.status,
-      title: title,
-      titleTranslations: releaseData.titleTranslations,
-      description: description,
-      descriptionTranslations: releaseData.descriptionTranslations,
-      releaseNote: releaseNote,
-      releaseNoteTranslations: releaseNoteTranslations,
-      publishDateUtc: releaseData.publishDateUtc,
-      canIgnoreRelease: releaseData.canIgnoreRelease,
-      reminderPeriod: releaseData.reminderPeriod,
-      releaseDelay: releaseData.releaseDelay,
-      stores: releaseData.stores,
-      customData: releaseData.customData,
-    );
-  }
 
   @override
   Release copyWith({
