+++ app_update_checker/lib/src/localizer/models/update_config.dart
